add crdt test
